+++ src/Ares/RuleFactory.php
@@ -19,6 +19,7 @@ use Ares\Rule\DirectoryRule;
 use Ares\Rule\EmailRule;
 use Ares\Rule\FileRule;
 use Ares\Rule\ForbiddenRule;
+use Ares\Rule\LengthRule;
 use Ares\Rule\MaxLengthRule;
 use Ares\Rule\MaxRule;
 use Ares\Rule\MinLengthRule;
@@ -45,6 +46,7 @@ class RuleFactory
         EmailRule::ID     => ['className' => EmailRule::class,     'reserved' => false],
         FileRule::ID      => ['className' => FileRule::class,      'reserved' => false],
         ForbiddenRule::ID => ['className' => ForbiddenRule::class, 'reserved' => false],
+        LengthRule::ID    => ['className' => LengthRule::class,    'reserved' => false],
         MaxLengthRule::ID => ['className' => MaxLengthRule::class, 'reserved' => false],
         MaxRule::ID       => ['className' => MaxRule::class,       'reserved' => false],
         MinLengthRule::ID => ['className' => MinLengthRule::class, 'reserved' => false],