+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improves validation schema parsing to detect inapplicable validation rule usage
 - adds ```tuple``` type to enable validation of fixed length heterogeneous array structures
 - adds ```numeric``` type to enable validation of numeric values (integer or float)
+- adds ```length``` validation rule to check the exact length of a string
 
 ## [v1.0.0] 2019-04-14
 ### Added